wip basecoin refactoring
