+++ examples/basecoin/cmd/basecoind/main.go
@@ -1,10 +1,13 @@
 package main
 
-import "github.com/cosmos/cosmos-sdk/examples/basecoin/app"
+import (
+	"github.com/cosmos/cosmos-sdk/baseapp"
+	"github.com/cosmos/cosmos-sdk/examples/basecoin/app"
+)
 
 func main() {
 	// TODO CREATE CLI
 
 	bapp := app.NewBasecoinApp("")
-	bapp.RunForever()
+	baseapp.RunForever(bapp)
 }
